[expo-updates] JS access to client log entries
Why

Once we have client-side logging in expo-updates, we need a way to expose it in JS so that developers can take log information and view it or send it to their own logging services.

How

New methods have been defined in the `Updates` module, and implemented in both the iOS and Android `Updates` native modules.

```ts
/**
* Retrieves the most recent expo-updates log entries from the client.
 *
 * @param maxAge Sets the max age of retrieved log entries in milliseconds. Default to 3600000 ms (1 hour).
 *
 * @return A promise that fulfills with an array of [`UpdatesLogEntry`](#updateslogentry) objects;
 *
 * The promise rejects if there is an unexpected error in retrieving the logs.
 */
Updates.readLogEntriesAsync(maxAge?: number): Promise<UpdatesLogEntry[]>;

/**
 * Clears existing expo-updates logs from the client.
 *
 * @return A promise that fulfills if the clear operation was successful.
 *
 * The promise rejects if there is an unexpected error in clearing the logs.
 *
 */
Updates.clearLogEntriesAsync(): Promise<void>;

/**
 * An object representing a single log entry from expo-updates logging on the client.
 */
type UpdatesLogEntry = {
  timestamp: number; // ms since 1/1/1970 UTC
  message: string;
  code: UpdatesLogEntryCode;
  level: string; // fatal, error, warn, info, debug, trace
  updateId?: string;
  assetId?: string;
  stacktrace?: string[];
};

enum UpdatesLogEntryCode {
  None = 'None',
  NoUpdatesAvailable = 'NoUpdatesAvailable',
  UpdateAssetsNotAvailable = 'UpdateAssetsNotAvailable',
  UpdateServerUnreachable = 'UpdateServerUnreachable',
  UpdateHasInvalidSignature = 'UpdateHasInvalidSignature',
  UpdateFailedToLoad = 'UpdateFailedToLoad',
  AssetsFailedToLoad = 'AssetsFailedToLoad',
  JSRuntimeError = 'JSRuntimeError'
}
```

Test Plan

A few log lines have been added on both iOS and Android so that logs will be generated for retrieval, and a sample app was used to verify functionality of the new methods.

There are also new Jest tests in `Updates-test.ios.ts`.